Refactor (move BatchedCubicSumcheck impls into dense_interleaved_poly and sparse_interleaved_poly)
