+++ jolt-core/src/poly/sparse_interleaved_poly.rs
@@ -1,5 +1,15 @@
-use super::{dense_interleaved_poly::DenseInterleavedPolynomial, dense_mlpoly::DensePolynomial};
-use crate::field::JoltField;
+use super::{
+    dense_interleaved_poly::DenseInterleavedPolynomial, dense_mlpoly::DensePolynomial,
+    split_eq_poly::SplitEqPolynomial, unipoly::UniPoly,
+};
+use crate::{
+    field::{JoltField, OptimizedMul},
+    subprotocols::{
+        grand_product::BatchedGrandProductLayer,
+        sumcheck::{BatchedCubicSumcheck, Bindable},
+    },
+    utils::{math::Math, thread::unsafe_allocate_zero_vec},
+};
 use rayon::prelude::*;
 
 #[derive(Default, Debug, Clone, Copy, PartialEq)]
@@ -82,6 +92,15 @@ impl<F: JoltField> SparseInterleavedPolynomial<F> {
         }
     }
 
+    pub fn coalesce(&self) -> Vec<F> {
+        let mut coalesced = vec![F::one(); self.dense_len];
+        self.coeffs
+            .iter()
+            .flatten()
+            .for_each(|sparse_coeff| coalesced[sparse_coeff.index] = sparse_coeff.value);
+        coalesced
+    }
+
     #[cfg(test)]
     pub fn interleave(left: &Vec<F>, right: &Vec<F>, batch_size: usize) -> Self {
         use itertools::Itertools;
@@ -152,8 +171,84 @@ impl<F: JoltField> SparseInterleavedPolynomial<F> {
             .flat_map(|segment| segment.par_chunk_by(|x, y| x.index / 4 == y.index / 4))
     }
 
+    pub fn layer_output(&self) -> Self {
+        if let Some(coalesced) = &self.coalesced {
+            let mut output = unsafe_allocate_zero_vec(coalesced.len().next_multiple_of(4) / 2);
+            output
+                .par_chunks_mut(2)
+                .zip(coalesced.par_chunks(4))
+                .for_each(|(output_chunk, input_chunk)| {
+                    let input_chunk = [
+                        *input_chunk.get(0).unwrap_or(&F::zero()),
+                        *input_chunk.get(1).unwrap_or(&F::zero()),
+                        *input_chunk.get(2).unwrap_or(&F::zero()),
+                        *input_chunk.get(3).unwrap_or(&F::zero()),
+                    ];
+                    output_chunk[0] = input_chunk[0] * input_chunk[2];
+                    output_chunk[1] = input_chunk[1] * input_chunk[3];
+                });
+
+            Self {
+                dense_len: self.dense_len / 2,
+                coeffs: vec![vec![]; self.batch_size()],
+                coalesced: Some(DenseInterleavedPolynomial::new(output)),
+            }
+        } else {
+            let coeffs: Vec<Vec<_>> = self
+                .coeffs
+                .par_iter()
+                .map(|segment| {
+                    segment
+                        .par_chunk_by(move |x, y| x.index / 4 == y.index / 4)
+                        .flat_map_iter(|sparse_block| {
+                            let mut outputs: Vec<SparseCoefficient<F>> = vec![];
+
+                            let block_index = sparse_block[0].index / 4;
+                            let mut dense_block = [F::one(); 4];
+                            for coeff in sparse_block {
+                                dense_block[coeff.index % 4] = coeff.value;
+                            }
+
+                            let left = dense_block[0].mul_1_optimized(dense_block[2]);
+                            let right = dense_block[1].mul_1_optimized(dense_block[3]);
+                            if !left.is_one() {
+                                let left_index = 2 * block_index;
+                                outputs.push((left_index, left).into());
+                            }
+                            if !right.is_one() {
+                                let right_index = 2 * block_index + 1;
+                                outputs.push((right_index, right).into());
+                            }
+
+                            outputs
+                        })
+                        .collect()
+                })
+                .collect();
+
+            Self::new(coeffs, self.dense_len / 2)
+        }
+    }
+}
+
+impl<F: JoltField> Bindable<F> for SparseInterleavedPolynomial<F> {
+    /// Incrementally binds a variable of this batched layer's polynomials.
+    /// If `self` is dense, we bind as in `BatchedDenseGrandProductLayer`,
+    /// processing nodes 4 at a time to preserve the interleaved order:
+    ///   0'  1'     2'  3'
+    ///   |\ |\      |\ |\
+    ///   | \| \     | \| \
+    ///   |  \  \    |  \  \
+    ///   |  |\  \   |  |\  \
+    ///   0  1 2  3  4  5 6  7
+    /// Left nodes have even indices, right nodes have odd indices.
+    /// If `self` is sparse, we basically do the same thing but with more
+    /// cases to check 😬
     #[tracing::instrument(skip_all, name = "SparseInterleavedPolynomial::bind")]
-    pub fn bind(&mut self, r: F) {
+    fn bind(&mut self, r: F) {
+        #[cfg(test)]
+        let (mut left_before_binding, mut right_before_binding) = self.uninterleave();
+
         if let Some(coalesced) = &mut self.coalesced {
             let padded_len = self.dense_len.next_multiple_of(4);
             coalesced.bind(r);
@@ -267,13 +362,235 @@ impl<F: JoltField> SparseInterleavedPolynomial<F> {
         self.dense_len /= 2;
         if (self.dense_len / self.batch_size()) == 2 {
             // Coalesce
-            let mut coalesced = vec![F::one(); self.dense_len];
-            self.coeffs
-                .iter()
-                .flatten()
-                .for_each(|sparse_coeff| coalesced[sparse_coeff.index] = sparse_coeff.value);
-            self.coalesced = Some(DenseInterleavedPolynomial::new(coalesced));
+            self.coalesced = Some(DenseInterleavedPolynomial::new(self.coalesce()));
         }
+
+        #[cfg(test)]
+        {
+            let (left_after_binding, right_after_binding) = self.uninterleave();
+            bind_left_and_right(&mut left_before_binding, &mut right_before_binding, r);
+
+            assert_eq!(
+                *self,
+                Self::interleave(
+                    &left_before_binding,
+                    &right_before_binding,
+                    self.batch_size()
+                )
+            );
+            assert_eq!(left_after_binding, left_before_binding);
+            assert_eq!(right_after_binding, right_before_binding);
+        }
+    }
+}
+
+impl<F: JoltField> BatchedGrandProductLayer<F> for SparseInterleavedPolynomial<F> {}
+impl<F: JoltField> BatchedCubicSumcheck<F> for SparseInterleavedPolynomial<F> {
+    #[cfg(test)]
+    fn sumcheck_sanity_check(&self, eq_poly: &SplitEqPolynomial<F>, round_claim: F) {
+        let merged_eq = eq_poly.merge();
+        let (left, right) = self.uninterleave();
+        let expected: F = left
+            .iter()
+            .zip(right.iter())
+            .zip(merged_eq.evals_ref().iter())
+            .map(|((l, r), eq)| *eq * l * r)
+            .sum();
+        assert_eq!(expected, round_claim);
+    }
+
+    /// We want to compute the evaluations of the following univariate cubic polynomial at
+    /// points {0, 1, 2, 3}:
+    ///     Σ eq(r, x) * left(x) * right(x)
+    /// where the inner summation is over all but the "least significant bit" of the multilinear
+    /// polynomials `eq`, `left`, and `right`. We denote this "least significant" variable x_b.
+    ///
+    /// Computing these evaluations requires processing pairs of adjacent coefficients of
+    /// `eq`, `left`, and `right`.
+    /// If `self` is dense, we process each layer 4 values at a time:
+    ///                  layer = [L, R, L, R, L, R, ...]
+    ///                           |  |  |  |
+    ///    left(0, 0, 0, ..., x_b=0) |  |  right(0, 0, 0, ..., x_b=1)
+    ///     right(0, 0, 0, ..., x_b=0)  left(0, 0, 0, ..., x_b=1)
+    /// If `self` is sparse, we basically do the same thing but with some fancy optimizations and
+    /// more cases to check 😬
+    #[tracing::instrument(skip_all, name = "BatchedSparseGrandProductLayer::compute_cubic")]
+    fn compute_cubic(&self, eq_poly: &SplitEqPolynomial<F>, previous_round_claim: F) -> UniPoly<F> {
+        if let Some(coalesced) = &self.coalesced {
+            return coalesced.compute_cubic(eq_poly, previous_round_claim);
+        }
+
+        let cubic_evals = if eq_poly.E1_len == 1 {
+            println!("E1_len == 1, not coaleseced");
+            let eq_evals: Vec<(F, F, F)> = eq_poly
+                .E2
+                .par_chunks(2)
+                .map(|eq_chunk| {
+                    let eval_point_0 = eq_chunk[0];
+                    let m_eq = eq_chunk[1] - eq_chunk[0];
+                    let eval_point_2 = eq_chunk[1] + m_eq;
+                    let eval_point_3 = eval_point_2 + m_eq;
+                    (eval_point_0, eval_point_2, eval_point_3)
+                })
+                .collect();
+            // TODO(moodlezoup): Can more efficiently compute these
+            let eq_eval_sums: (F, F, F) = eq_evals
+                .par_iter()
+                .fold(
+                    || (F::zero(), F::zero(), F::zero()),
+                    |sum, evals| (sum.0 + evals.0, sum.1 + evals.1, sum.2 + evals.2),
+                )
+                .reduce(
+                    || (F::zero(), F::zero(), F::zero()),
+                    |sum, evals| (sum.0 + evals.0, sum.1 + evals.1, sum.2 + evals.2),
+                );
+
+            let deltas: Vec<(F, F, F)> = self
+                .coeffs
+                .par_iter()
+                .flat_map(|segment| {
+                    segment
+                        .par_chunk_by(|x, y| x.index / 4 == y.index / 4)
+                        .map(|sparse_block| {
+                            let block_index = sparse_block[0].index / 4;
+                            let mut block = [F::one(); 4];
+                            for coeff in sparse_block {
+                                block[coeff.index % 4] = coeff.value;
+                            }
+
+                            let left = (block[0], block[2]);
+                            let right = (block[1], block[3]);
+
+                            let m_left = left.1 - left.0;
+                            let m_right = right.1 - right.0;
+
+                            let left_eval_2 = left.1 + m_left;
+                            let left_eval_3 = left_eval_2 + m_left;
+
+                            let right_eval_2 = right.1 + m_right;
+                            let right_eval_3 = right_eval_2 + m_right;
+
+                            let E2_eval = eq_poly.E2[block_index];
+                            // TODO(moodlezoup): Can save a multiplication here
+                            (
+                                E2_eval * (left.0 * right.0 - F::one()),
+                                E2_eval * (left_eval_2 * right_eval_2 - F::one()),
+                                E2_eval * (left_eval_3 * right_eval_3 - F::one()),
+                            )
+                        })
+                })
+                .collect();
+
+            deltas.into_par_iter().reduce(
+                || eq_eval_sums,
+                |sum, evals| (sum.0 + evals.0, sum.1 + evals.1, sum.2 + evals.2),
+            )
+        } else {
+            println!("E1_len != 1, not coalesced");
+            let deltas: Vec<(F, F, F, usize)> = self
+                .coeffs
+                .par_iter()
+                .flat_map(|segment| {
+                    segment
+                        .par_chunk_by(|x, y| x.index / 4 == y.index / 4)
+                        .map(|sparse_block| {
+                            let block_index = sparse_block[0].index / 4;
+                            let mut block = [F::one(); 4];
+                            for coeff in sparse_block {
+                                block[coeff.index % 4] = coeff.value;
+                            }
+
+                            let left = (block[0], block[2]);
+                            let right = (block[1], block[3]);
+
+                            let m_left = left.1 - left.0;
+                            let m_right = right.1 - right.0;
+
+                            let left_eval_2 = left.1 + m_left;
+                            let left_eval_3 = left_eval_2 + m_left;
+
+                            let right_eval_2 = right.1 + m_right;
+                            let right_eval_3 = right_eval_2 + m_right;
+
+                            let num_x1_bits = eq_poly.E1_len.log_2() - 1;
+                            let x1_bitmask = (1 << num_x1_bits) - 1;
+                            let x1 = block_index & x1_bitmask;
+                            let x2 = block_index >> num_x1_bits;
+
+                            let E2_eval = eq_poly.E2[x2];
+                            // TODO(moodlezoup): Can save a multiplication here
+                            (
+                                E2_eval * (left.0 * right.0 - F::one()),
+                                E2_eval * (left_eval_2 * right_eval_2 - F::one()),
+                                E2_eval * (left_eval_3 * right_eval_3 - F::one()),
+                                x1,
+                            )
+                        })
+                })
+                .collect();
+
+            let mut inner_sums: Vec<(F, F, F)> =
+                vec![(F::one(), F::one(), F::one()); eq_poly.E1_len / 2];
+            for delta in deltas.iter() {
+                let x1 = delta.3;
+                inner_sums[x1].0 += delta.0;
+                inner_sums[x1].1 += delta.1;
+                inner_sums[x1].2 += delta.2;
+            }
+
+            // Correct for the fact that the batch size is padded to a power of two
+            // with all-0 circuits.
+            // TODO(moodlezoup): optimize this
+            for x in (self.dense_len..self.dense_len.next_power_of_two()).step_by(4) {
+                let block_index = x / 4;
+                let num_x1_bits = eq_poly.E1_len.log_2() - 1;
+                let x1_bitmask = (1 << num_x1_bits) - 1;
+                let x1 = block_index & x1_bitmask;
+                let x2 = block_index >> num_x1_bits;
+                let E2_eval = eq_poly.E2[x2];
+
+                inner_sums[x1].0 -= E2_eval;
+                inner_sums[x1].1 -= E2_eval;
+                inner_sums[x1].2 -= E2_eval;
+            }
+
+            eq_poly.E1[..eq_poly.E1_len]
+                .par_chunks(2)
+                .zip(inner_sums.par_iter())
+                .map(|(E1_chunk, inner_sum)| {
+                    let E1_evals = {
+                        let eval_point_0 = E1_chunk[0];
+                        let m_eq = E1_chunk[1] - E1_chunk[0];
+                        let eval_point_2 = E1_chunk[1] + m_eq;
+                        let eval_point_3 = eval_point_2 + m_eq;
+                        (eval_point_0, eval_point_2, eval_point_3)
+                    };
+                    (
+                        E1_evals.0 * inner_sum.0,
+                        E1_evals.1 * inner_sum.1,
+                        E1_evals.2 * inner_sum.2,
+                    )
+                })
+                .reduce(
+                    || (F::zero(), F::zero(), F::zero()),
+                    |sum, evals| (sum.0 + evals.0, sum.1 + evals.1, sum.2 + evals.2),
+                )
+        };
+
+        let cubic_evals = [
+            cubic_evals.0,
+            previous_round_claim - cubic_evals.0,
+            cubic_evals.1,
+            cubic_evals.2,
+        ];
+
+        UniPoly::from_evals(&cubic_evals)
+    }
+
+    fn final_claims(&self) -> (F, F) {
+        assert_eq!(self.dense_len, 2);
+        let dense = self.to_dense();
+        (dense[0], dense[1])
     }
 }
