+++ jolt-core/src/subprotocols/grand_product.rs
@@ -3,9 +3,9 @@ use super::sumcheck::{BatchedCubicSumcheck, SumcheckInstanceProof};
 use crate::field::JoltField;
 use crate::poly::commitment::commitment_scheme::CommitmentScheme;
 use crate::poly::dense_interleaved_poly::DenseInterleavedPolynomial;
+use crate::poly::dense_mlpoly::DensePolynomial;
 use crate::poly::opening_proof::{ProverOpeningAccumulator, VerifierOpeningAccumulator};
 use crate::poly::split_eq_poly::SplitEqPolynomial;
-use crate::poly::{dense_mlpoly::DensePolynomial, unipoly::UniPoly};
 use crate::utils::math::Math;
 use crate::utils::thread::drop_in_background_thread;
 use crate::utils::transcript::ProofTranscript;
@@ -216,186 +216,6 @@ pub trait BatchedGrandProductLayer<F: JoltField>:
     }
 }
 
-/// Represents a single layer of a batched grand product circuit.
-/// A layer is assumed to be arranged in "interleaved" order, i.e. the natural
-/// order in the visual representation of the circuit:
-///      Λ        Λ        Λ        Λ
-///     / \      / \      / \      / \
-///   L0   R0  L1   R1  L2   R2  L3   R3   <- This is layer would be represented as [L0, R0, L1, R1, L2, R2, L3, R3]
-///                                           (as opposed to e.g. [L0, L1, L2, L3, R0, R1, R2, R3])
-#[derive(Debug, Clone)]
-pub struct BatchedDenseGrandProductLayer<F: JoltField> {
-    pub(crate) values: DenseInterleavedPolynomial<F>,
-}
-
-impl<F: JoltField> BatchedDenseGrandProductLayer<F> {
-    pub fn new(values: Vec<F>) -> Self {
-        Self {
-            values: DenseInterleavedPolynomial::new(values),
-        }
-    }
-}
-
-impl<F: JoltField> BatchedGrandProductLayer<F> for BatchedDenseGrandProductLayer<F> {}
-impl<F: JoltField> BatchedCubicSumcheck<F> for BatchedDenseGrandProductLayer<F> {
-    #[cfg(test)]
-    fn sumcheck_sanity_check(&self, eq_poly: &SplitEqPolynomial<F>, round_claim: F) {
-        let (left, right) = self.values.uninterleave();
-        let merged_eq = eq_poly.merge();
-        let expected: F = left
-            .iter()
-            .zip(right.iter())
-            .zip(merged_eq.evals_ref().iter())
-            .map(|((l, r), eq)| *eq * l * r)
-            .sum();
-        assert_eq!(expected, round_claim);
-    }
-
-    /// Incrementally binds a variable of this batched layer's polynomials.
-    /// Even though each layer is backed by a single Vec<F>, it represents two polynomials
-    /// one for the left nodes in the circuit, one for the right nodes in the circuit.
-    /// These two polynomials' coefficients are interleaved into one Vec<F>. To preserve
-    /// this interleaved order, we bind values like this:
-    ///   0'  1'     2'  3'
-    ///   |\ |\      |\ |\
-    ///   | \| \     | \| \
-    ///   |  \  \    |  \  \
-    ///   |  |\  \   |  |\  \
-    ///   0  1 2  3  4  5 6  7
-    /// Left nodes have even indices, right nodes have odd indices.
-    #[tracing::instrument(skip_all, name = "BatchedDenseGrandProductLayer::bind")]
-    fn bind(&mut self, eq_poly: &mut SplitEqPolynomial<F>, r: &F) {
-        let _ = rayon::join(|| self.values.bind(*r), || eq_poly.bind(*r));
-    }
-
-    /// We want to compute the evaluations of the following univariate cubic polynomial at
-    /// points {0, 1, 2, 3}:
-    ///     Σ eq(r, x) * left(x) * right(x)
-    /// where the inner summation is over all but the "least significant bit" of the multilinear
-    /// polynomials `eq`, `left`, and `right`. We denote this "least significant" variable x_b.
-    ///
-    /// Computing these evaluations requires processing pairs of adjacent coefficients of
-    /// `eq`, `left`, and `right`.
-    /// Recall that the `left` and `right` polynomials are interleaved in each layer of `self.layers`,
-    /// so we process each layer 4 values at a time:
-    ///                  layer = [L, R, L, R, L, R, ...]
-    ///                           |  |  |  |
-    ///    left(0, 0, 0, ..., x_b=0) |  |  right(0, 0, 0, ..., x_b=1)
-    ///     right(0, 0, 0, ..., x_b=0)  left(0, 0, 0, ..., x_b=1)
-    #[tracing::instrument(skip_all, name = "BatchedDenseGrandProductLayer::compute_cubic")]
-    fn compute_cubic(&self, eq_poly: &SplitEqPolynomial<F>, previous_round_claim: F) -> UniPoly<F> {
-        // debug_assert_eq!(self.values.len(), 2 * eq_poly.len());
-
-        let cubic_evals = if eq_poly.E1_len == 1 {
-            self.values
-                .par_chunks(4)
-                .zip(eq_poly.E2.par_chunks(2))
-                .map(|(layer_chunk, eq_chunk)| {
-                    let eq_evals = {
-                        let eval_point_0 = eq_chunk[0];
-                        let m_eq = eq_chunk[1] - eq_chunk[0];
-                        let eval_point_2 = eq_chunk[1] + m_eq;
-                        let eval_point_3 = eval_point_2 + m_eq;
-                        (eval_point_0, eval_point_2, eval_point_3)
-                    };
-                    let left = (
-                        *layer_chunk.get(0).unwrap_or(&F::zero()),
-                        *layer_chunk.get(2).unwrap_or(&F::zero()),
-                    );
-                    let right = (
-                        *layer_chunk.get(1).unwrap_or(&F::zero()),
-                        *layer_chunk.get(3).unwrap_or(&F::zero()),
-                    );
-
-                    let m_left = left.1 - left.0;
-                    let m_right = right.1 - right.0;
-
-                    let left_eval_2 = left.1 + m_left;
-                    let left_eval_3 = left_eval_2 + m_left;
-
-                    let right_eval_2 = right.1 + m_right;
-                    let right_eval_3 = right_eval_2 + m_right;
-
-                    (
-                        eq_evals.0 * left.0 * right.0,
-                        eq_evals.1 * left_eval_2 * right_eval_2,
-                        eq_evals.2 * left_eval_3 * right_eval_3,
-                    )
-                })
-                .reduce(
-                    || (F::zero(), F::zero(), F::zero()),
-                    |sum, evals| (sum.0 + evals.0, sum.1 + evals.1, sum.2 + evals.2),
-                )
-        } else {
-            let num_E1_chunks = eq_poly.E1_len / 2;
-
-            let mut evals = (F::zero(), F::zero(), F::zero());
-            for (x1, E1_chunk) in eq_poly.E1[..eq_poly.E1_len].chunks(2).enumerate() {
-                let E1_evals = {
-                    let eval_point_0 = E1_chunk[0];
-                    let m_eq = E1_chunk[1] - E1_chunk[0];
-                    let eval_point_2 = E1_chunk[1] + m_eq;
-                    let eval_point_3 = eval_point_2 + m_eq;
-                    (eval_point_0, eval_point_2, eval_point_3)
-                };
-                let inner_sums = eq_poly.E2[..eq_poly.E2_len]
-                    .par_iter()
-                    .zip(self.values.par_chunks(4).skip(x1).step_by(num_E1_chunks))
-                    .map(|(E2_eval, P_x1)| {
-                        let left = (
-                            *P_x1.get(0).unwrap_or(&F::zero()),
-                            *P_x1.get(2).unwrap_or(&F::zero()),
-                        );
-                        let right = (
-                            *P_x1.get(1).unwrap_or(&F::zero()),
-                            *P_x1.get(3).unwrap_or(&F::zero()),
-                        );
-
-                        let m_left = left.1 - left.0;
-                        let m_right = right.1 - right.0;
-
-                        let left_eval_2 = left.1 + m_left;
-                        let left_eval_3 = left_eval_2 + m_left;
-
-                        let right_eval_2 = right.1 + m_right;
-                        let right_eval_3 = right_eval_2 + m_right;
-
-                        // TODO(moodlezoup): can save a mult by E2_eval here
-                        (
-                            *E2_eval * left.0 * right.0,
-                            *E2_eval * left_eval_2 * right_eval_2,
-                            *E2_eval * left_eval_3 * right_eval_3,
-                        )
-                    })
-                    .reduce(
-                        || (F::zero(), F::zero(), F::zero()),
-                        |sum, evals| (sum.0 + evals.0, sum.1 + evals.1, sum.2 + evals.2),
-                    );
-
-                evals.0 += E1_evals.0 * inner_sums.0;
-                evals.1 += E1_evals.1 * inner_sums.1;
-                evals.2 += E1_evals.2 * inner_sums.2;
-            }
-            evals
-        };
-
-        let cubic_evals = [
-            cubic_evals.0,
-            previous_round_claim - cubic_evals.0,
-            cubic_evals.1,
-            cubic_evals.2,
-        ];
-        UniPoly::from_evals(&cubic_evals)
-    }
-
-    fn final_claims(&self) -> (F, F) {
-        assert_eq!(self.values.len(), 2);
-        let left_claim = self.values.coeffs[0];
-        let right_claim = self.values.coeffs[1];
-        (left_claim, right_claim)
-    }
-}
-
 /// A batched grand product circuit.
 /// Note that the circuit roots are not included in `self.layers`
 ///        o
@@ -405,7 +225,7 @@ impl<F: JoltField> BatchedCubicSumcheck<F> for BatchedDenseGrandProductLayer<F> 
 ///   o   o o   o  <- layers[layers.len() - 2]
 ///       ...
 pub struct BatchedDenseGrandProduct<F: JoltField> {
-    layers: Vec<BatchedDenseGrandProductLayer<F>>,
+    layers: Vec<DenseInterleavedPolynomial<F>>,
 }
 
 impl<F: JoltField, PCS: CommitmentScheme<Field = F>> BatchedGrandProduct<F, PCS>
@@ -422,17 +242,16 @@ impl<F: JoltField, PCS: CommitmentScheme<Field = F>> BatchedGrandProduct<F, PCS>
         assert!((leaves.len() / batch_size).is_power_of_two());
 
         let num_layers = (leaves.len() / batch_size).log_2();
-        let mut layers: Vec<BatchedDenseGrandProductLayer<F>> = Vec::with_capacity(num_layers);
-        layers.push(BatchedDenseGrandProductLayer::new(leaves));
+        let mut layers: Vec<DenseInterleavedPolynomial<F>> = Vec::with_capacity(num_layers);
+        layers.push(DenseInterleavedPolynomial::new(leaves));
 
         for i in 0..num_layers - 1 {
             let previous_layer = &layers[i];
             let new_layer = previous_layer
-                .values
                 .par_chunks(2)
                 .map(|chunk| chunk[0] * chunk[1])
                 .collect();
-            layers.push(BatchedDenseGrandProductLayer::new(new_layer));
+            layers.push(DenseInterleavedPolynomial::new(new_layer));
         }
 
         Self { layers }
@@ -449,7 +268,6 @@ impl<F: JoltField, PCS: CommitmentScheme<Field = F>> BatchedGrandProduct<F, PCS>
     fn claimed_outputs(&self) -> Vec<F> {
         let last_layer = &self.layers[self.layers.len() - 1];
         last_layer
-            .values
             .par_chunks(2)
             .map(|chunk| chunk[0] * chunk[1])
             .collect()
@@ -466,8 +284,9 @@ impl<F: JoltField, PCS: CommitmentScheme<Field = F>> BatchedGrandProduct<F, PCS>
 #[cfg(test)]
 mod tests {
     use super::*;
-    use crate::poly::{
-        commitment::zeromorph::Zeromorph, dense_interleaved_poly::bind_left_and_right,
+    use crate::{
+        poly::{commitment::zeromorph::Zeromorph, dense_interleaved_poly::bind_left_and_right},
+        subprotocols::sumcheck::Bindable,
     };
     use ark_bn254::{Bn254, Fr};
     use ark_std::test_rng;
@@ -507,10 +326,7 @@ mod tests {
             >>::construct((leaves.concat(), batch_size));
 
             for layer in &batched_circuit.layers {
-                assert_eq!(
-                    layer.values.coeffs.par_iter().product::<Fr>(),
-                    expected_product
-                );
+                assert_eq!(layer.coeffs.par_iter().product::<Fr>(), expected_product);
             }
 
             let claimed_outputs: Vec<Fr> = <BatchedDenseGrandProduct<Fr> as BatchedGrandProduct<
@@ -541,21 +357,14 @@ mod tests {
             .take(batch_size)
             .collect();
 
-            let mut layer = BatchedDenseGrandProductLayer::<Fr>::new(values.concat());
+            let mut layer = DenseInterleavedPolynomial::<Fr>::new(values.concat());
+            let (mut expected_left_poly, mut expected_right_poly) = layer.uninterleave();
 
-            let mut eq_poly = SplitEqPolynomial::new(
-                &std::iter::repeat_with(|| Fr::random(&mut rng))
-                    .take(batch_size.next_power_of_two().log_2())
-                    .collect::<Vec<Fr>>(),
-            );
             let r = Fr::random(&mut rng);
-
-            let (mut expected_left_poly, mut expected_right_poly) = layer.values.uninterleave();
-
-            layer.bind(&mut eq_poly, &r);
+            layer.bind(r);
             bind_left_and_right(&mut expected_left_poly, &mut expected_right_poly, r);
 
-            let (actual_left_poly, actual_right_poly) = layer.values.uninterleave();
+            let (actual_left_poly, actual_right_poly) = layer.uninterleave();
             assert_eq!(expected_left_poly, actual_left_poly);
             assert_eq!(expected_right_poly, actual_right_poly);
         }
